+++ src/test/kotlin/toolkit/monkeyTest/NextFrequencyBin.kt
@@ -6,6 +6,6 @@ import kotlin.random.Random
 fun nextFrequencyBin(): FrequencyBin {
     return FrequencyBin(
         frequency = Random.nextFloat(),
-        amplitude = Random.nextDouble()
+        amplitude = Random.nextFloat()
     )
 }
